pyproject.toml
[tool.poetry]
name = "tegracli"
version = "0.1.0"
description = ""
authors = ["Philipp Kessling"]

[tool.poetry.dependencies]
python = "^3.9"
Telethon = "^1.24.0"
ujson = "^5.3.0"
click = "^8.1.3"
loguru = "^0.6.0"
PyYAML = "^6.0"
pandas = "^1.4.3"

[tool.poetry.dev-dependencies]
ipykernel = "^6.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tegracli = 'tegracli.main:cli'